feat: refactored add artboard logic
